moved carouselUtil into carousel
